pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repovore"
version = "0.1.0"
description = "GitLab repository quality assessment pipeline: fetch, enrich, score, search"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Lavi Shpigelman"},
]
keywords = ["github", "repository", "quality", "pipeline", "assessment"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "PyGithub>=2.0",
    "requests>=2.28",
    "rich>=13.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.25"]
web = ["fastapi>=0.115", "uvicorn[standard]>=0.34", "jinja2>=3.1"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.0",
    "ruff>=0.4.0",
    "mypy>=1.0",
    "coverage>=7.0",
]

[project.scripts]
repovore = "repovore.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/repovore"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "types-requests>=2.33.0.20260402",
]
